Start merging API and APIProxy classes
